Fix client and expenses model, adding and priotize administration tab

+++ TrackExpenses/Controllers/ExpensesController.cs
@@ -1,7 +1,11 @@
-﻿using TrackExpenses.App_Start;
-using TrackExpenses.Models;
-using Microsoft.AspNetCore.Http;
+﻿using TrackExpenses.Models;
 using Microsoft.AspNetCore.Mvc;
+using Microsoft.AspNetCore.Identity;
+using TrackExpenses.Data;
+using Microsoft.EntityFrameworkCore;
+using Microsoft.AspNetCore.Mvc.Rendering;
+
+//Controller expenses
 
 namespace TrackExpenses.Controllers
 {
@@ -9,39 +13,44 @@ namespace TrackExpenses.Controllers
     {
         // GET: Expenses
         private readonly FinancasDbContext _context;
+        private readonly UserManager<Client> _userManager;
 
-        public ExpensesController(FinancasDbContext context)
+
+        public ExpensesController(FinancasDbContext context, UserManager<Client> userManager)
         {
             _context = context;
+            _userManager = userManager;
+
         }
 
+        [HttpGet]
         public IActionResult ListExpenses()
         {
             //List ALL expenses
-            var allExpenses  = _context.Expenses.ToList();
-            if (allExpenses != null)
-                {
-                return View(allExpenses);
-            }
-            else
-            {
-                return View();
-            }
-        }
-        public IActionResult CreateGroupClient()
-        {
-            return View();
-        }
-            public IActionResult CreateEditExpense(int? id)
-        {
+            var user = _context?.Clients.FirstOrDefault(userToFind => userToFind.Email == User.Identity.Name);
+            if (user == null) return View();
+            
+            var client = _context?.Clients.Include(client => client.Expenses).FirstOrDefault(x => x.Email == user.Email);
+            if (client == null) return View();
 
+            return View(client.Expenses);
+
+        }
+
+
+        public IActionResult CreateEditExpense(int? id)
+        {
+           
             if (id != null) 
             {
                 //editing  -> load an expense by Id
                 var expenseInDB = _context.Expenses.SingleOrDefault(expense => expense.Id == id);
+                
                 return View(expenseInDB);
 
             }
+            var categories = _context.ExpenseCategory.ToList();
+            ViewBag.Categories = new SelectList(categories, "Id", "Name");
             return View();
         }
 
@@ -49,26 +58,42 @@ namespace TrackExpenses.Controllers
         {
             //Deleat expense by Id
             var expenseInDB = _context.Expenses.SingleOrDefault(expense => expense.Id == id);
+            if (expenseInDB == null) return View();
             _context.Expenses.Remove(expenseInDB);
             _context.SaveChanges();
             return RedirectToAction("ListExpenses");
         }
 
-        public IActionResult CreateEditExpenseForm(Expense model)
+        
+        [HttpPost]
+        public async Task<IActionResult> CreateEditExpenseForm(Expense model)
         {
+ 
             if(model.Id == 0)
             {
                 //Create
+                
+
+                var user = await _userManager.GetUserAsync(User);
+                
+                if (user == null) return RedirectToAction("ListExpenses");
+                
+                var client = _context.Clients.FirstOrDefault(x => x.Email == user.Email);
+                if (client == null) return RedirectToAction("ListExpenses");
+                
+                model.ClientId = client.Id;
+                model.GroupId = client.GroupId;
+                client.Expenses.Add(model);
                 _context.Expenses.Add(model);
-            }
-            else{
-                //Editing
-                _context.Expenses.Update(model);
+                _context.Clients.Update(client);
 
             }
-            _context.SaveChanges();
-            return RedirectToAction("ListExpenses");
+
+                await _context.SaveChangesAsync();
+                return RedirectToAction("ListExpenses");
+
+            
         }
-
+        
     }
 }
